quickstarts: fix: Arguments class was moved (#1969)
