+++ quickstarts/kit/dynamic-docker-image-file-multi-layer/pom.xml
@@ -20,7 +20,7 @@
 
   <groupId>org.eclipse.jkube.quickstarts.kit</groupId>
   <artifactId>dynamic-docker-image-file-multi-layer</artifactId>
-  <version>1.12.0</version>
+  <version>1.13-SNAPSHOT</version>
   <name>Eclipse JKube :: Quickstarts :: Kit :: Docker Image from Multilayer Dockerfile</name>
   <packaging>jar</packaging>
