+++ quickstarts/kit/dynamic-docker-image-file-multi-layer/src/main/java/org/eclipse/jkube/quickstart/kit/docker/dynamic/DynamicDockerfileGenerator.java
@@ -14,10 +14,10 @@
 package org.eclipse.jkube.quickstart.kit.docker.dynamic;
 
 import org.apache.commons.io.FileUtils;
+import org.eclipse.jkube.kit.common.Arguments;
 import org.eclipse.jkube.kit.common.AssemblyConfiguration;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.image.ImageConfiguration;
-import org.eclipse.jkube.kit.config.image.build.Arguments;
 import org.eclipse.jkube.kit.config.image.build.BuildConfiguration;
 import org.eclipse.jkube.kit.config.image.build.DockerFileBuilder;
